+++ app/main.wasp
@@ -78,40 +78,10 @@ page PlayGroundPage {
   component: import PlayGroundPageWithCustomAuth from "@src/client/app/PlayGroundPage"
 }
 
-route BuildRoute { path: "/build/:id?", to: BuildPage }
-page BuildPage {
-  component: import BuildPageWithCustomAuth from "@src/client/app/BuildPage"
-}
-
 route BuildRouteNew { path: "/build-new/:type?", to: BuildPageNew }
 page BuildPageNew {
   component: import BuildPageNewWithCustomAuth from "@src/client/app/BuildPageNew"
 }
-
-// route SecretsRoute { path: "/build/secrets", to: SecretsPage }
-// page SecretsPage {
-//   component: import BuildPageWithCustomAuth from "@src/client/app/BuildPage"
-// }
-
-// route ToolBoxRoute { path: "/build/toolboxes", to: ToolBoxPage }
-// page ToolBoxPage {
-//   component: import BuildPageWithCustomAuth from "@src/client/app/BuildPage"
-// }
-
-// route LLMsRoute { path: "/build/llms", to: LLMsPage }
-// page LLMsPage {
-//   component: import BuildPageWithCustomAuth from "@src/client/app/BuildPage"
-// }
-
-// route TeamsRoute { path: "/build/teams", to: TeamsPage }
-// page TeamsPage {
-//   component: import BuildPageWithCustomAuth from "@src/client/app/BuildPage"
-// }
-
-// route AgentsRoute { path: "/build/agents", to: AgentsPage }
-// page AgentsPage {
-//   component: import BuildPageWithCustomAuth from "@src/client/app/BuildPage"
-// }
 
 route LoginRoute { path: "/login", to: LoginPage }
 page LoginPage {
