+++ app/src/client/components/SelectTeamToChat.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect, useRef } from 'react';
 import { useHistory } from 'react-router-dom';
 import { type Chat } from 'wasp/entities';
-import { SelectedModelSchema } from '../interfaces/BuildPageInterfaces';
+import { SelectedModelSchema } from '../interfaces/BuildPageInterfacesNew';
 import { CreateNewChatProps } from '../interfaces/PlaygroundPageInterface';
 import CustomBreadcrumb from './CustomBreadcrumb';
 import NotificationBox from './NotificationBox';
