--- app/src/client/components/buildPage/useDetectRefresh.ts
@@ -1,17 +1,0 @@
-/* This file should be deleted */
-
-import { useEffect } from 'react';
-
-const useDetectRefresh = (onRefresh: () => void) => {
-  useEffect(() => {
-    if (sessionStorage.getItem('isLoaded')) {
-      // This is a refresh
-      onRefresh();
-    } else {
-      // This is the initial load
-      sessionStorage.setItem('isLoaded', 'true');
-    }
-  }, []);
-};
-
-export default useDetectRefresh;
