+++ app/src/client/utils/constants.ts
@@ -62,6 +62,3 @@ Before you begin, ensure you have the following:
 <span class="ml-10"> while creating the Fly.io API Token in the deployment steps below.</span>
 </div>
 `;
-
-export const SELECT_PLACEHOLDER = '--- Select Placeholder ---';
-export const SELECT_CLEAR_PLACEHOLDER = '--- Selected Option Cleared ---';
